+++ util/suffixes/n2n/case_suffixes.py
@@ -1,5 +1,5 @@
 from util.suffix import Suffix, Type, HasMajorHarmony, HasMinorHarmony, SuffixGroup
-
+## COMPLETE
 class CaseSuffix(Suffix):
     def __init__(self, name, suffix, 
                 comes_to=Type.NOUN,
