+++ util/suffixes/n2v/verbifiers.py
@@ -1,18 +1,17 @@
 from util.suffix import Suffix, Type, HasMajorHarmony, HasMinorHarmony, SuffixGroup
-import util.word_methods as wrd
-
 ## COMPLETE
-class Plural(Suffix):
+class VerbifierSuffix(Suffix):
     def __init__(self, name, suffix, 
-                comes_to=Type.NOUN,
-                makes=Type.NOUN,
-                 major_harmony=None, 
+                 comes_to=Type.NOUN,
+                 makes=Type.VERB,
+                 major_harmony=HasMajorHarmony.Yes, 
                  minor_harmony=None, 
                  needs_y_buffer=False, 
                  form_function=None,
-                 group=SuffixGroup.PLURAL, 
+                 group=SuffixGroup.N2V_DERIVATIONAL, 
                  is_unique=False):
         
+                # Dynamic default assignment for minor harmony
         if minor_harmony is None:
             # If the suffix contains any narrow vowel, it defaults to having minor harmony
             if any(vowel in suffix for vowel in ['ı', 'i', 'u', 'ü']): # only i is enough bc of the standart narrow front vowel converntion
@@ -20,17 +19,12 @@ class Plural(Suffix):
             else:
                 minor_harmony = HasMinorHarmony.No
 
-        if needs_y_buffer is None:
-            if suffix[0] in ['a', 'e', 'ı', 'i', 'o', 'ö', 'u', 'ü']:
-                needs_y_buffer = True
-            else:
-                needs_y_buffer = False
         super().__init__(
             name=name,
             suffix=suffix,
             comes_to=comes_to,
             makes=makes,
-            form_function=form_function ,
+            form_function=form_function, 
             major_harmony=major_harmony,
             minor_harmony=minor_harmony,
             needs_y_buffer=needs_y_buffer,
@@ -39,13 +33,14 @@ class Plural(Suffix):
         )
 
 
+absentative_se = VerbifierSuffix("absentative_se", "se")
+onomatopea_de  = VerbifierSuffix("onomatopea_de",  "de")
+verbifier_e    = VerbifierSuffix("verbifier_e",    "e" )
+aplicative_le  = VerbifierSuffix("aplicative_le",  "le")
+verbifier_ik   = VerbifierSuffix("verbifier_ik",   "ik")
 
 
-
-
-plural_ler  = Plural("plural_ler", "ler")
-
-PLURALS = [
+VERBIFIERS = [ 
     value for name, value in globals().items() 
     if isinstance(value, Suffix) and name != "Suffix"
 ]
